backend/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Image Processing
Pillow==11.0.0

# Database
sqlalchemy==2.0.46
aiosqlite==0.22.1

# Utilities
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
